pyproject.toml
[project]
name = "ai-employee"
version = "0.4.0"
description = "Personal AI Employee - Platinum Tier implementation"
requires-python = ">=3.13"
dependencies = [
    # Bronze
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    # Silver — Gmail
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.0.0",
    # Silver — Scheduling
    "schedule>=1.2.0",
    # Silver — MCP server
    "mcp>=1.0.0",
    # Silver — Email via SMTP
    "aiosmtplib>=3.0.0",
    # Gold — Odoo ERP
    "httpx>=0.27.0",
    # Gold — Structured logging
    "structlog>=24.0.0",
    # Silver — WhatsApp Business webhook server
    "flask>=3.0.0",
]

[project.scripts]
ai-employee      = "orchestrator:main"
file-watcher     = "watchers.filesystem_watcher:main"
gmail-watcher    = "watchers.gmail_watcher:main"
linkedin-watcher   = "watchers.linkedin_watcher:main"
whatsapp-watcher   = "watchers.whatsapp_watcher:main"
social-watcher   = "watchers.social_watcher:main"
scheduler        = "scheduler:main"
email-mcp        = "mcp_servers.email_mcp_server:main"
odoo-mcp         = "mcp_servers.odoo_mcp_server:main"
social-mcp       = "mcp_servers.social_mcp_server:main"
audit-mcp        = "mcp_servers.audit_mcp_server:main"
# Platinum Tier
cloud-agent      = "cloud_agent:main"
health-monitor   = "cloud.health_monitor:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["watchers", "mcp_servers"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
